Implement basic search index generation
This implements a basic search index. For each XML document processed
that has an Atom <entry> element as the root element, the text content
of the <title> element is broken into words, converted to uppercase,
and recorded in a map associating terms with the URIs of processed
documents.

Known limitations: The <entry> elements that appear inside feeds
are essentially ignored. The reason for this is that we can't assign
unique URIs for individual feed entries, so we depend on those entries
providing links to documents containing individual entries. Thankfully,
most OPDS feeds already appear to do this.

Fix: https://github.com/AULFA/opdsget/issues/10